4.JPaint/src/test/java/LetterTests.java
import org.junit.Assert;
import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.Matchers;
import org.mockito.Mock;
import org.mockito.Mockito;
import org.mockito.runners.MockitoJUnitRunner;
import pl.edu.amu.bawsj.encog.recognition.FromImageNumberReader.Letter.BitImage;
import pl.edu.amu.bawsj.encog.recognition.FromImageNumberReader.Letter.BoardArray;
import pl.edu.amu.bawsj.encog.recognition.FromImageNumberReader.Letter.Letter;

/**
 * Created by JanJa on 12.02.2017.
 */

public class LetterTests {

    @Test
    public void LetterCreate() {
        int[][] arr = {{1, 1, 0}, {1, 1, 0}, {1, 1, 0}};
        BitImage bitImage = new BitImage(arr);
        Letter letter = new Letter(bitImage);
        Assert.assertNotNull(letter);
    }


    @Test
    public void removeRedundantLinesShouldWork() {


        int[][] arrWR = {{1, 1, 1}, {1, 1, 0}, {1, 1, 0}, {1, 1, 1}};
        int[][] arr = {{1, 1, 0}, {1, 1, 0}};
        BitImage bitImageWR = new BitImage(arrWR);
        Letter letterWR = new Letter(bitImageWR);


        BitImage bitImage = new BitImage(arr);
        Letter letter = new Letter(bitImage);

        System.out.println(letter.toString());
        System.out.println(letterWR.toString());
        Assert.assertEquals(letter.toString(), letterWR.toString());


    }


    @Test
    public void shouldScaleSmallerCorrectly() {
        int[][] arr = {{1, 0, 0, 0, 0},
                {1, 0, 0, 0, 0},
                {1, 0, 0, 0, 0},
                {1, 0, 0, 0, 0}
        };
        BitImage bitImage = new BitImage(arr);

        Letter letter = new Letter(bitImage);

        int[][] expected = {{1, 1, 0, 0, 0, 0, 0, 0, 0, 0},
                {1, 1, 0, 0, 0, 0, 0, 0, 0, 0},
                {1, 1, 0, 0, 0, 0, 0, 0, 0, 0},
                {1, 1, 0, 0, 0, 0, 0, 0, 0, 0},
                {1, 1, 0, 0, 0, 0, 0, 0, 0, 0},
                {1, 1, 0, 0, 0, 0, 0, 0, 0, 0},
                {1, 1, 0, 0, 0, 0, 0, 0, 0, 0},
                {1, 1, 0, 0, 0, 0, 0, 0, 0, 0},
                {1, 1, 0, 0, 0, 0, 0, 0, 0, 0},
                {1, 1, 0, 0, 0, 0, 0, 0, 0, 0}
        };
        BoardArray boardArr = new BoardArray(expected);
        Assert.assertTrue(boardArr.toString().equals(letter.toString()));


    }

    @Test
    public void shouldScaleBiggerCorrectly() {
        int[][] arr = {{0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1},
                {1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1},
                {0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1},
                {1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1},
                {0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1},
                {1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1},
                {0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1},
                {1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1},
                {0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1},
                {1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1},
                {0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1},
                {1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1},
                {0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1},
                {1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1},
                {0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1},
        };
        BitImage bitImage = new BitImage(arr);

        Letter letter = new Letter(bitImage);

        int[][] expected = {{0, 0, 1, 1, 1, 1, 1, 1, 1, 1},
                {0, 0, 1, 1, 1, 1, 1, 1, 1, 1},
                {0, 0, 1, 1, 1, 1, 1, 1, 1, 1},
                {0, 0, 1, 1, 1, 1, 1, 1, 1, 1},
                {0, 0, 1, 1, 1, 1, 1, 1, 1, 1},
                {0, 0, 1, 1, 1, 1, 1, 1, 1, 1},
                {0, 0, 1, 1, 1, 1, 1, 1, 1, 1},
                {0, 0, 1, 1, 1, 1, 1, 1, 1, 1},
                {0, 0, 1, 1, 1, 1, 1, 1, 1, 1},
                {0, 0, 1, 1, 1, 1, 1, 1, 1, 1}
        };
        BoardArray boardArr = new BoardArray(expected);

        Assert.assertTrue(boardArr.toString().equals(letter.toString()));


    }



    @Test
    public void LettersShouldTranslateToDoubleArray() {
        int[][] arr = {{1, 1, 0}, {1, 1, 0}, {1, 1, 0}};
        BitImage bitImage = new BitImage(arr);
        Letter letter = new Letter(bitImage);

        double[] doubleArr = letter.toDoubleArray();
double[] exp = {1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,0.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,0.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,0.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,0.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,0.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,0.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,0.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,0.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,0.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,0.0};



      Assert.assertArrayEquals(exp,doubleArr,0.0);

    }


}
